refactor

+++ webapp/src/main/java/by/epam/tc/conference/web/controller/command/impl/proposal/DeleteProposalCommand.java
@@ -1,8 +1,9 @@
 package by.epam.tc.conference.web.controller.command.impl.proposal;
 
+import by.epam.tc.conference.entity.Proposal;
+import by.epam.tc.conference.entity.UserPrincipal;
 import by.epam.tc.conference.services.ProposalService;
-import by.epam.tc.conference.services.ServiceException;
-import by.epam.tc.conference.web.controller.command.CommandException;
+import by.epam.tc.conference.services.exception.ServiceException;
 import by.epam.tc.conference.web.controller.command.impl.AbstractCommand;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -11,7 +12,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public class DeleteProposalCommand extends AbstractCommand {
+
     private static final Logger logger = LogManager.getLogger(DeleteProposalCommand.class);
+    private static final String PROPOSAL_ID_PARAM = "id";
     private final ProposalService proposalService;
 
     public DeleteProposalCommand(ProposalService proposalService) {
@@ -19,13 +22,29 @@ public class DeleteProposalCommand extends AbstractCommand {
     }
 
     @Override
-    public String execute(HttpServletRequest request, HttpServletResponse response) throws CommandException {
-        String idString = request.getParameter("id");
-        Long id = Long.valueOf(idString);
+    public String execute(HttpServletRequest request, HttpServletResponse response) {
+        logger.traceEntry();
+        Long id = parseIdParameter(request, PROPOSAL_ID_PARAM);
+        if (id == null) {
+            logger.debug("Failed to delete proposal. Incorrect proposal id");
+            return processBadRequest(request, response);
+        }
         try {
-            proposalService.deleteProposal(id);
-            return "redirect:";
+            Proposal proposal = proposalService.getProposal(id);
+            long participantId = proposal.getParticipantId();
+            UserPrincipal user = getUser(request);
+            long userId = user.getId();
+            String query;
+            if (participantId == userId) {
+                proposalService.deleteProposal(id);
+                query = "redirect:/user-dashboard";
+            } else {
+                logger.debug("Failed to delete proposal id={}. Request forbidden. User id={} is not the owner", id);
+                query = forbidRequest(request, response);
+            }
+            return query;
         } catch (ServiceException e) {
+            logger.error("Failed to delete proposal id={}", id, e);
             return processInternalError(request, response);
         }
     }
