+++ webapp/src/main/java/by/epam/tc/conference/web/controller/command/impl/section/DeleteSectionCommand.java
@@ -1,14 +1,15 @@
 package by.epam.tc.conference.web.controller.command.impl.section;
 
 import by.epam.tc.conference.services.SectionService;
-import by.epam.tc.conference.services.ServiceException;
-import by.epam.tc.conference.web.controller.command.CommandException;
+import by.epam.tc.conference.services.exception.ServiceException;
 import by.epam.tc.conference.web.controller.command.impl.AbstractCommand;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public class DeleteSectionCommand extends AbstractCommand {
+
+    private static final String SECTION_ID_PARAM = "id";
     private final SectionService sectionService;
 
     public DeleteSectionCommand(SectionService sectionService) {
@@ -16,9 +17,11 @@ public class DeleteSectionCommand extends AbstractCommand {
     }
 
     @Override
-    public String execute(HttpServletRequest request, HttpServletResponse response) throws CommandException {
-        String idString = request.getParameter("id");
-        Long id = Long.valueOf(idString);
+    public String execute(HttpServletRequest request, HttpServletResponse response) {
+        Long id = parseIdParameter(request, SECTION_ID_PARAM);
+        if (id == null) {
+            return processBadRequest(request, response);
+        }
         try {
             sectionService.deleteSection(id);
             return "redirect:";
